Create relativistic_velocity
This function can be used to calculate the velocity of an object based on its energy and mass, assuming that the object is moving at relativistic speeds (i.e., speeds close to the speed of light). Specifically, the function calculates the velocity of the object based on the equation E = mc^2 / √(1 - v^2/c^2), which is a fundamental equation in the theory of relativity.

This function can be useful in various applications, such as in physics research or engineering design. For example, it could be used to calculate the velocity of high-energy particles in a particle accelerator, or to determine the velocity of an object that is being propelled by a high-energy laser. The function could also be used in simulations or models to predict the behavior of relativistic systems.

Overall, this function provides a convenient way to calculate the velocity of an object based on its energy and mass, and can be a valuable tool in various fields of science and engineering.